+++ internal/proposal/builder.go
