+++ internal/proposal/util.go
