+++ pkg/chaincode/endorser.go
@@ -6,9 +6,9 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/hyperledger/fabric-admin-sdk/internal/proposal"
 	"github.com/hyperledger/fabric-admin-sdk/internal/protoutil"
 	"github.com/hyperledger/fabric-admin-sdk/pkg/identity"
-	"github.com/hyperledger/fabric-admin-sdk/pkg/internal/proposal"
 	"github.com/hyperledger/fabric-protos-go-apiv2/common"
 	"github.com/hyperledger/fabric-protos-go-apiv2/orderer"
 	"github.com/hyperledger/fabric-protos-go-apiv2/peer"
