+++ pkg/chaincode/queryinstalled.go
@@ -9,8 +9,8 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/hyperledger/fabric-admin-sdk/internal/proposal"
 	"github.com/hyperledger/fabric-admin-sdk/pkg/identity"
-	"github.com/hyperledger/fabric-admin-sdk/pkg/internal/proposal"
 
 	"github.com/hyperledger/fabric-protos-go-apiv2/peer"
 	"github.com/hyperledger/fabric-protos-go-apiv2/peer/lifecycle"
