+++ pkg/channel/channel.go
@@ -10,9 +10,9 @@ import (
 	"net/http"
 
 	"github.com/hyperledger/fabric-admin-sdk/internal/osnadmin"
+	"github.com/hyperledger/fabric-admin-sdk/internal/proposal"
 	"github.com/hyperledger/fabric-admin-sdk/internal/protoutil"
 	"github.com/hyperledger/fabric-admin-sdk/pkg/identity"
-	"github.com/hyperledger/fabric-admin-sdk/pkg/internal/proposal"
 
 	cb "github.com/hyperledger/fabric-protos-go-apiv2/common"
 	pb "github.com/hyperledger/fabric-protos-go-apiv2/peer"
